flow control now available in the interface
There is still a latency issue a pressing "stop".

+++ src/main/java/org/jcompas/control/InfinitePlayer.java
@@ -24,7 +24,7 @@ package org.jcompas.control;
  * @author thibautd
  */
 public interface InfinitePlayer {
-	public void start(final long startTime, final int compasLengthMilli);
+	public void start(final long startTime, final long compasLengthMilli);
 	public void stop();
 }
